Change Orientation to "Button Orientation"
I changed the SendTelemetry for the button orientation to send "Button Orientation" instead of "Orientation". This will be less confusing for people who will assume "Orientation" might be related to the device orientation or possible gyroscope sensor in the device. "Button Orientation" is more descriptive of what event is actually being reported. :)