BT510 magnet data improvement
Added logic to always send the BT510 magnet state.  The magnet state is always present in the flags data and the magnet event (2) does not get sent from the BT510 each time it changes, so just read it out.  Additionally, if the magnet state changes, always send it up in a telemetry message right away.